requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
pydantic>=2.5.1
pydantic-settings>=2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==3.2.2
python-multipart>=0.0.6
python-dotenv>=1.0.0
aiohttp>=3.9.1
websockets>=12.0
slowapi==0.1.8
limits>=2.3
deprecated>=1.2
redis==5.0.1
aiosmtplib==2.0.2
stripe>=5.0.0
psycopg2-binary>=2.9.9
email-validator>=2.1.0
asyncpg>=0.28.0
gunicorn>=20.1.0
rx>=3.2.0
httpx>=0.27.0
psutil>=5.9.0
alembic>=1.7.5
jinja2>=3.1.2
orjson>=3.9.0
apscheduler>=3.10.0
pytz>=2023.3
anthropic>=0.21.0
tenacity>=8.2.3